splitlab::split_pattern
Trait Pattern

pub trait Pattern {
    // Required method
    fn find_next(&self, s: &str) -> Option<(usize, usize)>;
}

An interface for a type that can find itself in a string.

Required Methods

fn find_next(&self, s: &str) -> Option<(usize, usize)>

Finds the next occurrence of the pattern in the given string.

Returns Some((start, end)) if the pattern is found, where start is starting index of the pattern and end is the index of the end of the pattern in the string. Returns None if the pattern is not found.